Fix group_authority representation in repr
The current `__repr__` implementation of `Aarc_g002_entitlement` is
assuming that the `group_authority` is always defined and present.
However, this is not always the case. In the cases that the
`group_authority` is not defined `#None` will be displayed as a suffix
of the `repr` result. To fix this, we check whether `group_authority` is
defined and only if that condition is met we display it, exactly as is
already done with `role`.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>