Fix strict regex and is_contained_in/__le__
- subgroups are optional
- role is optional

The group might not be followed by any of the two - it should look out
for "#" along with ":".

Applying the same logic for the subgroups - the regex should look out
for "#" along with ":".

Finally, the role should not take over the group-authority, but should
look out for the first "#".

---

Fixing the regex relevealed a faulty test case:
- TestAarc_g002_entitlement::test_subgroup_required

Previously, `req_entitlement` would _not_ be recognized as an aarc-g002
entitlement. `is_contained_in` would be handled by the non-g002 handler.

Fixing the regex to recognize the req_entitlement as an g002-entitlement
revealed a bug in `is_contained_in`/`__le__`. The subgroups of self should
be checked, not the subgroups of other. All of the subgroups of
req_entitlement should also exist in other.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>